Fix bug in SAMDUSBEndpointImpl::send_data()
The USB peripheral requires that the `addr' field of the endpoint descriptor be
a pointer into RAM.  If the `buf' parameter of this function is a pointer to
data in flash, this causes bogus data to be transmitted to the host.  In the
case that `buf' does not point to RAM, just copy the contents into the already
allocated buffer pointed to by `desc->bank1.addr'.